Factor out DemoComponents
